requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.13
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-timezone-field==6.1.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
idna==3.7
kombu==5.3.7
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pillow==10.4.0
prompt-toolkit==3.0.47
psycopg2==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
